Backend/app/routers/stats.py
from fastapi import APIRouter, HTTPException
from pydantic import BaseModel
from typing import Any, Dict

router = APIRouter()

class StatsRequest(BaseModel):
    # Placeholder: define relevant fields as needed
    data: Dict[str, Any]

@router.post("/analyze")
def analyze_stats(request: StatsRequest):
    """
    Perform statistical analysis on provided data.
    """
    try:
        # TODO: implement statistical analysis logic
        results = {"status": "not implemented", "data": request.data}
        return results
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))
