Backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.5
pandas>=1.4.0
numpy>=1.23.0
scikit-learn>=1.1.0
scikit-image>=0.19.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
openpyxl>=3.0.0

